[UPDATE] Update CCCV2ShowCase due element changes on Pie chart.

+++ src/test/java/com/pentaho/ctools/utils/RunnableFindElementInvisible.java
@@ -14,55 +14,53 @@ import com.google.common.base.Function;
 
 public class RunnableFindElementInvisible implements Runnable {
 
-  private WebDriver driver;
-  private Integer timeout;
-  private Integer pollingTime;
-  private By locator;
+	private WebDriver driver;
+	private long timeout;
+	private long pollingTime;
+	private By locator;
 
-  private WebElement theElement;
+	private WebElement theElement;
 
-  public RunnableFindElementInvisible( WebDriver driver, Integer timeout, Integer pollingTime, By locator ) {
-    super();
-    this.driver = driver;
-    this.timeout = timeout;
-    this.pollingTime = pollingTime;
-    this.locator = locator;
-  }
+	public RunnableFindElementInvisible( WebDriver driver, long timeout, long pollingTime, By locator ) {
+		super();
+		this.driver = driver;
+		this.timeout = timeout;
+		this.pollingTime = pollingTime;
+		this.locator = locator;
+	}
 
-  public WebElement getInvisbleElement() {
-    return this.theElement;
-  }
+	public WebElement getInvisbleElement() {
+		return this.theElement;
+	}
 
-  public By getLocator() {
-    return this.locator;
-  }
+	public By getLocator() {
+		return this.locator;
+	}
 
-  @Override
-  public void run() {
-    Wait<WebDriver> wait =
-        new FluentWait<>( this.driver ).withTimeout( this.timeout, TimeUnit.SECONDS ).pollingEvery( this.pollingTime,
-            TimeUnit.MILLISECONDS );
+	@Override
+	public void run() {
+		Wait<WebDriver> wait = new FluentWait<>( this.driver ).withTimeout( this.timeout, TimeUnit.SECONDS ).pollingEvery( this.pollingTime, TimeUnit.MILLISECONDS );
 
-    // Wait for element present and not visible
-    this.theElement = wait.until( new Function<WebDriver, WebElement>() {
+		// Wait for element present and not visible
+		this.theElement = wait.until( new Function<WebDriver, WebElement>() {
 
-      @Override
-      public WebElement apply( WebDriver d ) {
-        try {
-          List<WebElement> listElem = d.findElements( getLocator() );
-          if ( listElem.size() > 0 ) {
-            WebElement elem = listElem.get( 0 );
-            if ( elem != null && ( elem.isEnabled() && !elem.isDisplayed() ) ) {
-              return elem;
-            }
-            return null;
-          }
-          return null;
-        } catch ( StaleElementReferenceException sere ) {
-          return null;
-        }
-      }
-    } );
-  }
+			@Override
+			public WebElement apply( WebDriver d ) {
+				try {
+					List<WebElement> listElem = d.findElements( getLocator() );
+					if ( listElem.size() > 0 ) {
+						WebElement elem = listElem.get( 0 );
+						if ( elem != null && ( elem.isEnabled() && !elem.isDisplayed() ) ) {
+							return elem;
+						}
+						return null;
+					}
+					return null;
+				} catch ( StaleElementReferenceException sere ) {
+					return null;
+				}
+			}
+		} );
+	}
 
 }
